Create README.md
As we know,COVID-19 has affected the whole world very badly. It has a huge impact on our everyday life, and this crisis is increasing day by day. In the near future, it seems difficult to eradicate this virus completely. To counter this virus, Face Masks have became an integral part of our Lives. So here we've developed a model that will be able to detect whether person is wearing Mask or not. This is Convolution Neural Network Based Classifier to classify people whether they are wearing mask or not.